Regression test which does have to use real network moved into a separate Test class.
This is the only test from ClientRegressionTest which does rely on real network.
