+++ hazelcast-client/src/test/java/com/hazelcast/client/ClientRegressionWithRealNetworkTest.java
@@ -1,0 +1,47 @@
+package com.hazelcast.client;
+
+import com.hazelcast.client.config.ClientConfig;
+import com.hazelcast.config.Config;
+import com.hazelcast.core.Hazelcast;
+import com.hazelcast.core.HazelcastInstance;
+import com.hazelcast.core.IMap;
+import com.hazelcast.test.HazelcastSerialClassRunner;
+import com.hazelcast.test.HazelcastTestSupport;
+import com.hazelcast.test.annotation.QuickTest;
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
+import org.junit.runner.RunWith;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
+
+@RunWith(HazelcastSerialClassRunner.class)
+@Category(QuickTest.class)
+public class ClientRegressionWithRealNetworkTest extends HazelcastTestSupport {
+
+    @Test
+    public void testClientPortConnection() {
+        final Config config1 = new Config();
+        config1.getGroupConfig().setName("foo");
+        config1.getNetworkConfig().setPort(5701);
+        final HazelcastInstance instance1 = Hazelcast.newHazelcastInstance(config1);
+        instance1.getMap("map").put("key", "value");
+
+        final Config config2 = new Config();
+        config2.getGroupConfig().setName("bar");
+        config2.getNetworkConfig().setPort(5702);
+        HazelcastInstance instance2 = Hazelcast.newHazelcastInstance(config2);
+
+        final ClientConfig clientConfig = new ClientConfig();
+        clientConfig.getGroupConfig().setName("bar");
+        final HazelcastInstance client = HazelcastClient.newHazelcastClient(clientConfig);
+
+        final IMap<Object, Object> map = client.getMap("map");
+        assertNull(map.put("key", "value"));
+        assertEquals(1, map.size());
+
+        client.shutdown();
+        instance1.getLifecycleService().terminate();
+        instance2.getLifecycleService().terminate();
+    }
+}
